+++ recipes/default.rb
@@ -66,15 +66,15 @@ cookbook_file "/tmp/zbx_templates_base_e42.xml" do
   group "root"
 end
 
-zabbix_client_template "/tmp/zbx_templates_base_e42.xml"
+zabbix_template "/tmp/zbx_templates_base_e42.xml"
 
-zabbix_client_host node.fqdn do
+zabbix_host node.fqdn do
   host_group "default"
   use_ip true
   ip_address "127.0.0.1"
 end
 
-zabbix_client_template "CPU_E42_Template" do
+zabbix_template "CPU_E42_Template" do
   action :add
   host_name node.fqdn
 end